+++ Modern JavaScript From The Beginning/Section 8: API Projects/78 - 81. WeatherJS/api.js
@@ -1,0 +1,19 @@
+class Country {
+    constructor (name) {
+        this.name = name;
+    }
+
+    async getCountryInfo () {
+        const response = await fetch(
+            `https://restcountries.eu/rest/v2/name/${encodeURIComponent(this.name)}`
+        );
+
+        const countryInfo = await response.json();
+
+        return countryInfo;
+    }
+
+    changeCountry (name) {
+        this.name = name;
+    }
+}