+++ Modern JavaScript From The Beginning/Section 8: API Projects/78 - 81. WeatherJS/app.js
@@ -1,0 +1,19 @@
+const defaultCountry = 'Norway';
+
+const countryAPI = new Country(defaultCountry);
+
+function getCountryInfo () {
+    countryAPI
+        .getCountryInfo()
+        .then(
+            data => console.log(data)
+        )
+        .catch(
+            error => console.log(error)
+        );
+}
+
+document.addEventListener(
+    'DOMContentLoaded',
+    getCountryInfo
+);